Add back the stack_string
