+++ include/util_str.h
@@ -5,6 +5,8 @@
 #include <iomanip>
 #include <string>
 
+#include "impl/util_str_impl.h"
+
 namespace small {
     // to lower
     // clang-format off
@@ -151,89 +153,17 @@ namespace small {
     //
     // string conversions
     //
-    namespace impl {
-        // from utf8 -> utf16 needed length
-        inline std::size_t to_utf16_needed_length(const char* mbstr /*src*/)
-        {
-            std::setlocale(LC_ALL, "en_US.utf8");
-            std::mbstate_t state{};
-
-            // determine size
-            std::size_t new_length = 0;
-#if defined(_WIN32) || defined(_WIN64)
-            int ret = mbsrtowcs_s(&new_length, nullptr, 0, &mbstr, 0, &state);
-            if (ret != 0 || new_length == 0)
-                return static_cast<std::size_t>(-1);
-            --new_length; // because it adds the null terminator in length
-#else
-            new_length = std::mbsrtowcs(nullptr, &mbstr, 0, &state);
-            if (new_length == static_cast<std::size_t>(-1))
-                return static_cast<std::size_t>(-1);
-#endif
-            return new_length;
-        }
-
-        // from utf16 -> utf8
-        inline std::size_t to_utf8_needed_length(const wchar_t* wstr /*src*/)
-        {
-            std::setlocale(LC_ALL, "en_US.utf8");
-            std::mbstate_t state{};
-
-            // determine size
-            std::size_t new_length = 0;
-
-#if defined(_WIN32) || defined(_WIN64)
-            int ret = wcsrtombs_s(&new_length, nullptr, 0, &wstr, 0, &state);
-            if (ret != 0 || new_length == 0)
-                return static_cast<std::size_t>(-1);
-            --new_length; // because it adds the null terminator in length
-#else
-            new_length = std::wcsrtombs(nullptr, &wstr, 0, &state);
-            if (new_length == static_cast<std::size_t>(-1))
-                return static_cast<std::size_t>(-1);
-#endif
-            return new_length;
-        }
-
-        // from utf8 -> utf16
-        inline void to_utf16(const char* mbstr /*src*/, wchar_t* wstr /*dest*/, const std::size_t& new_length /*dest lengt*/)
-        {
-            std::setlocale(LC_ALL, "en_US.utf8");
-            std::mbstate_t state{};
-
-#if defined(_WIN32) || defined(_WIN64)
-            size_t converted = 0;
-            /* ret              = */ mbsrtowcs_s(&converted, wstr, new_length + 1, &mbstr, new_length, &state);
-#else
-            /*size_t converted =*/std::mbsrtowcs(wstr, &mbstr, new_length, &state);
-#endif
-        }
-
-        // from utf16 -> utf8
-        inline void to_utf8(const wchar_t* wstr /*src*/, char* mbstr /*dest*/, const std::size_t& new_length /*dest length*/)
-        {
-            std::setlocale(LC_ALL, "en_US.utf8");
-            std::mbstate_t state{};
-
-#if defined(_WIN32) || defined(_WIN64)
-            size_t converted = 0;
-            /* ret              = */ wcsrtombs_s(&converted, mbstr, new_length + 1, &wstr, new_length, &state);
-#else
-            /*size_t converted =*/std::wcsrtombs(mbstr, &wstr, new_length, &state);
-#endif
-        }
-    } // namespace impl
 
     // from utf8 -> utf16
     inline std::wstring to_utf16(const char* mbstr)
     {
         std::wstring wstr;
-        std::size_t  new_length = small::impl::to_utf16_needed_length(mbstr);
+        std::size_t  new_length = small::strimpl::to_utf16_needed_length(mbstr);
         if (new_length == static_cast<std::size_t>(-1))
             return wstr;
 
         wstr.resize(new_length);
-        small::impl::to_utf16(mbstr, wstr.data(), wstr.size());
+        small::strimpl::to_utf16(mbstr, wstr.data(), wstr.size());
         return wstr;
     }
 
@@ -241,12 +171,12 @@ namespace small {
     inline std::string to_utf8(const wchar_t* wstr)
     {
         std::string str;
-        std::size_t new_length = small::impl::to_utf8_needed_length(wstr);
+        std::size_t new_length = small::strimpl::to_utf8_needed_length(wstr);
         if (new_length == static_cast<std::size_t>(-1))
             return str;
 
         str.resize(new_length);
-        small::impl::to_utf8(wstr, str.data(), str.size());
+        small::strimpl::to_utf8(wstr, str.data(), str.size());
         return str;
     }
 } // namespace small
